+++ semcore/side-panel/src/SidePanel.jsx
@@ -11,7 +11,7 @@ import keyboardFocusEnhance from '@semcore/utils/lib/enhances/keyboardFocusEnhan
 import usePreventScroll from '@semcore/utils/lib/use/usePreventScroll';
 import { Text } from '@semcore/typography';
 import ArrowLeft from '@semcore/icon/ArrowLeft/m';
-import Button from '@semcore/button';
+import Button, { ButtonLink } from '@semcore/button';
 import { cssVariableEnhance } from '@semcore/utils/lib/useCssVariable';
 import { useFocusLock } from '@semcore/utils/lib/use/useFocusLock';
 import { useContextTheme } from '@semcore/utils/lib/ThemeProvider';
@@ -233,17 +233,12 @@ function Title(props) {
 }
 
 function Back(props) {
-  const SBack = Root;
-  const SBackText = Button.Text;
   const { Children, styles } = props;
 
   return sstyled(styles)(
-    <SBack render={Button} use='tertiary' theme='muted' type='button'>
-      <Button.Addon tag={ArrowLeft} />
-      <SBackText>
-        <Children />
-      </SBackText>
-    </SBack>,
+    <ButtonLink color={'text-hint'} size={100} addonLeft={ArrowLeft}>
+      <Children />
+    </ButtonLink>,
   );
 }