+++ semcore/button/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.36.0] - 2024-08-29
+
+### Added
+
+- Component `ButtonLink`.
+
 ## [5.35.2] - 2024-08-23
 
 ### Changed