+++ semcore/card/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.41.0] - 2024-08-29
+
+### Changed
+
+- Render hints in header with `ButtonLink` component.
+
 ## [5.40.4] - 2024-08-23
 
 ### Changed